Query cache:
React-Query references each query with its queryKey and queryFunction and caches the return data for a default time of 5min.
When you launch the same query within that time, react-query keeps isLoading as false, displays the cached data and in the meanwhile fetches data in the background to seemlessly update the UI without displaying redundant loading state in the event server state has been updated.
You can track that background fetching with isFeching from useQuery hook, or pass it a third parameter object to change default cache time.