Polling
Polling refers to the process of fetching data at regular intervals.
It enables you to keep the UI in sync irrespective of configurations like
refetchOnMount or refetchOnWindowFocus which relies on user interactions.
refetchInterval set the polling interval and setting refetchIntervalInBackground to true allow polling even out of focus providing a pleasant user experience for example for stock market app.